Checkbox component rewritten to TS
