+++ src/components/Checkbox/Checkbox.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Checkbox from '@components/Checkbox'
+import { Checkbox } from '@components/Checkbox'
 
 # Checkbox
